ash/requirements.txt
flask==2.3.3
flask-cors==4.0.0
qdrant-client==1.6.4
transformers==4.35.2
torch==2.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
datasets==2.14.6
joblib==1.3.2
sentence-transformers==2.2.2
werkzeug==2.3.7